+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Improved suggestion for pipeline operator in malformed query (tolerant mode)
+- Added suggestion for object keys in array-like notation property accessor
 
 ## 1.0.0-beta.2 (May 17, 2020)